requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.4.0
attrs==25.3.0
backoff==2.2.1
base58==2.1.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bech32==1.2.0
bitarray==2.9.3
blinker==1.9.0
browserbase==1.2.0
cached-property==2.0.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==1.0.2
click==8.1.8
contourpy==1.3.1
cosmpy==0.9.2
coverage==7.8.0
cryptography==44.0.2
cycler==0.12.1
cytoolz==1.0.1
deprecation==2.1.0
distro==1.9.0
docker==6.1.2
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.16.1
entrypoints==0.4
eth-abi==5.2.0
eth-account==0.8.0
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.2
faker==37.1.0
favicon==0.7.0
flask==2.1.3
fonttools==4.56.0
frozenlist==1.5.0
gitdb==4.0.12
gitpython==3.1.44
googleapis-common-protos==1.69.2
gotrue==2.12.0
gql==3.5.0
graphql-core==3.2.6
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
hexbytes==0.3.1
hpack==4.1.0
htbuilder==0.9.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
ipfshttpclient==0.8.0a2
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.3.3
jwt==1.3.1
kiwisolver==1.4.8
langchain==0.3.22
langchain-core==0.3.50
langchain-openai==0.3.12
langchain-text-splitters==0.3.7
langsmith==0.3.23
lru-dict==1.2.0
lxml==5.3.1
markdown==3.7
markdownlit==0.0.7
markupsafe==3.0.2
matplotlib==3.10.1
mcp==1.6.0
mech-client==0.4.0
morphys==1.0
multiaddr==0.0.9
multidict==6.2.0
narwhals==1.32.0
netaddr==1.3.0
numpy==2.2.4
open-aea==1.65.0
open-aea-cli-ipfs==1.65.0
open-aea-ledger-cosmos==1.65.0
open-aea-ledger-ethereum==1.65.0
open-autonomy==0.19.7
openai==1.70.0
orjson==3.10.16
packaging==23.2
pandas==2.2.3
paramiko==3.5.1
parsimonious==0.10.0
pillow==11.1.0
playwright==1.51.0
plotly==6.0.1
pluggy==1.5.0
postgrest==1.0.1
prometheus-client==0.21.1
propcache==0.3.1
protobuf==4.24.4
py-ecc==8.0.0b2
py-evm==0.10.1b1
py-multibase==1.0.3
py-multicodec==0.2.1
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.1
pydantic-core==2.33.0
pydantic-settings==2.8.1
pydeck==0.9.1
pyee==12.1.1
pyjwt==2.10.1
pymdown-extensions==10.14.3
pymultihash==0.8.2
pynacl==1.5.0
pyparsing==3.2.3
pyrsistent==0.20.0
pytest==7.2.1
pytest-mock==3.14.0
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2025.2
pyunormalize==16.0.0
pyyaml==6.0.1
realtime==2.4.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rlp==3.0.0
safe-eth-py==6.4.0
safe-pysha3==1.0.4
semver==2.13.0
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlalchemy==2.0.40
sse-starlette==2.2.1
st-annotated-text==4.0.2
st-theme==1.2.3
starlette==0.46.1
storage3==0.11.3
streamlit==1.44.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
strenum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tabulate==0.9.0
tenacity==9.0.0
texttable==1.6.7
tiktoken==0.9.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
trie==3.1.0
typing-extensions==4.13.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
uuid==1.30
uvicorn==0.34.0
validators==0.34.0
valory-docker-compose==1.29.3
varint==1.0.2
watchdog==6.0.0
web3==6.20.4
websocket-client==0.59.0
websockets==13.1
werkzeug==2.0.3
yarl==1.18.3
zstandard==0.23.0
browser-use==0.1.41
boto3